LayoutTests/platform/glib/imported/w3c/web-platform-tests/cors/304-expected.txt

CONSOLE MESSAGE: Origin http://web-platform.test:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://not-web-platform.test:8800/cors/resources/304.py?id=7&r=%s1659396386435 due to access control checks.
CORS - 304 Responses


PASS A 304 response with no CORS headers inherits from the stored response
PASS A 304 can expand Access-Control-Expose-Headers
PASS A 304 can contract Access-Control-Expose-Headers
PASS A 304 can change Access-Control-Allow-Origin
